refactor: 💡 make contextual routing hook more generic
